backend/cookie_extractor_backend/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
chromedriver-autoinstaller==0.4.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
h11==0.13.0
idna==3.3
outcome==1.2.0
PySocks==1.7.1
pytz==2022.1
requests==2.28.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wsproto==1.2.0
